e_tenders_india/src/requirements.txt
beautifulsoup4==4.9.3
boto3==1.17.98
botocore==1.20.98
certifi==2021.5.30
chardet==4.0.0
idna==2.10
jmespath==0.10.0
lxml==4.6.3
numpy==1.21.0
openpyxl==3.0.7
pandas==1.2.5
python-dateutil==2.8.1
python-dotenv==0.17.1
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
s3transfer==0.4.2
selenium==3.141.0
six==1.16.0
soupsieve==2.2.1
tqdm==4.61.1
urllib3==1.26.5
xlrd==2.0.1
elasticsearch==7.13.4
opensearch-py==1.0.0
requests-aws4auth==1.1.1
